sql: separate pgcrypto_builtins into their own file
Release note: None
