+++ pkg/sql/sem/builtins/pg_builtins.go
@@ -110,7 +110,7 @@ func PGIOBuiltinPrefix(typ *types.T) string {
 	return builtinPrefix
 }
 
-// initPGBuiltins adds all of the postgres builtins to the Builtins map.
+// initPGBuiltins adds all of the postgres builtins to the builtins map.
 func initPGBuiltins() {
 	for k, v := range pgBuiltins {
 		if _, exists := builtins[k]; exists {
