+++ pkg/sql/sem/builtins/window_builtins.go
@@ -21,7 +21,7 @@ import (
 )
 
 func initWindowBuiltins() {
-	// Add all windows to the Builtins map after a few sanity checks.
+	// Add all windows to the builtins map after a few sanity checks.
 	for k, v := range windows {
 		if _, exists := builtins[k]; exists {
 			panic("duplicate builtin: " + k)
